ETL Pipeline/requirement.txt
pandas
SQLAlchemy
psycopg2-binary
pyyaml
